Move php_hrtime_current() to zend_monotonic_time()
