+++ UPGRADING.INTERNALS
@@ -67,6 +67,9 @@ PHP 8.3 INTERNALS UPGRADE NOTES
   - zend_set_user_opcode_handler
   - zend_ssa_inference
 * Removed unused macros PHP_FNV1_32A_INIT and PHP_FNV1A_64_INIT. See GH-11114.
+* Removed ext/standard/hrtime.h, PHP_HRTIME_ macros, HRTIME_AVAILABLE, and
+  php_hrtime_current().
+* Added Zend/zend_time.h, ZEND_MONOTONIC_TIME_AVAILABLE, zend_monotonic_time()
 
 ========================
 2. Build system changes
