+++ configure.ac
@@ -396,6 +396,7 @@ dirent.h \
 sys/param.h \
 sys/types.h \
 sys/time.h \
+time.h \
 netinet/in.h \
 alloca.h \
 arpa/inet.h \
@@ -1734,6 +1735,7 @@ PHP_ADD_SOURCES(Zend, \
     zend_default_classes.c zend_inheritance.c zend_smart_str.c zend_cpuinfo.c zend_gdb.c \
     zend_observer.c zend_system_id.c zend_enum.c zend_fibers.c zend_atomic.c \
     zend_max_execution_timer.c \
+    zend_time.c \
 	Optimizer/zend_optimizer.c \
 	Optimizer/pass1.c \
 	Optimizer/pass3.c \
