+++ ext/standard/php_standard.h
@@ -23,7 +23,6 @@
 #include "php_mail.h"
 #include "md5.h"
 #include "sha1.h"
-#include "hrtime.h"
 #include "html.h"
 #include "exec.h"
 #include "file.h"
